Include unistd.h to avoid warnings about read and write calls. (ttstat): Make it work on x86_64.
FossilOrigin-Name: ea8cb5f15fa1dee7d14102553d5242df7e6f0af9b9c2c66e6183f093c2d12bc2

+++ tty/termcap/ttyio.c
@@ -21,6 +21,7 @@
 #endif
 
 #include	<errno.h>
+#include	<unistd.h>
 #include	<sys/ioctl.h>
 #ifdef sun
 #include	<sys/filio.h>
@@ -175,7 +176,7 @@ ttclose (void)
 int
 ttstat (void)
 {
-  long n;
+  int n;
 
   if (ioctl (0, FIONREAD, &n) < 0)
     return (FALSE);
